notifications and kyc

+++ backend/resources/kyc.py
@@ -1,15 +1,15 @@
 import werkzeug
-from flask import request
+from flask import request, url_for
 from flask_jwt_extended import current_user, jwt_required
 from flask_restplus import Resource, fields, abort, Namespace
 from flask_restplus.reqparse import RequestParser
 
-from backend.ext import pagination, db
 from backend.common import phone_number
 from backend.common.schema import KYCSchema
+from backend.ext import pagination, db
 from backend.models import Kyc
 from backend.resources.users import user_schema
-from backend.utils import delete_file, file_upload
+from backend.utils import delete_file, file_upload, roles_required
 
 parser = RequestParser(bundle_errors=True, trim=True)
 
@@ -35,48 +35,40 @@ mschema = ns_kyc.model('KYC', {
 
 
 class KYCResourceList(Resource):
-    method_decorators = [jwt_required]
-    """
-    This is for administrators
-    """
+    method_decorators = [roles_required(['ADMIN', 'USER']), jwt_required]
 
     def get(self):
         return pagination.paginate(Kyc, schema, True)
 
 
 class KYCResource(Resource):
-    method_decorators = [jwt_required]
+    method_decorators = [roles_required(['ADMIN', 'USER']), jwt_required]
 
     @ns_kyc.marshal_with(mschema)
     def get(self, pk):
         obj = Kyc.query.get_or_404(pk)
         return obj
 
-    # @api.marshal_with(mschema)
     def put(self, pk):
-        put_parser = RequestParser()
+        put_parser = RequestParser(trim=True, bundle_errors=True)
         put_parser.add_argument('status', required=True, location='json', type=str,
                                 choices=['Approved', 'Pending', 'Disapproved', 'Suspended', 'Processing'])
         args = put_parser.parse_args(strict=True)
         kyc = Kyc.query.get_or_404(pk)
         kyc.status = args.status
-        kyc.save()
-        return kyc
+        return kyc.save(**args)
 
     def delete(self, pk):
         kyc = Kyc.query.get_or_404(pk)
         return kyc.delete(), 202
 
 
-##
-
 class ClientKYCResource(Resource):
-    method_decorators = [jwt_required]
-    mschema.pop('user')
+    method_decorators = [roles_required(['CLIENT']), jwt_required]
 
     @ns_kyc.marshal_with(mschema)
     def get(self):
-        return current_user.kyc_doc
+        return current_user.kyc_doc if current_user.kyc_doc else abort(404)
 
     def post(self):
         args = parser.parse_args(strict=True)
@@ -88,6 +80,8 @@ class ClientKYCResource(Resource):
 
     def put(self):
         if current_user.kyc_doc:
+            if current_user.kyc_doc.status != 'Pending':
+                return {'message': 'You cant update kyc when not in pending state'}, 400
             args = parser.parse_args(strict=True)
             kyc = schema.load(data=request.json, instance=current_user.kyc_doc, session=db.session, unknown='exclude')
             return kyc.save(**args), 200
@@ -97,19 +91,22 @@ class ClientKYCResource(Resource):
     def delete(self):
         doc = current_user.kyc_doc
         if doc:
+            if doc.status != 'Pending':
+                return {'message': 'You cant update kyc when not in pending state'}, 400
             delete_file(doc.file)
             return doc.delete(), 202
         return abort(404)
 
 
 class KycUploadResource(Resource):
-    method_decorators = [jwt_required]
+    method_decorators = [roles_required(['CLIENT']), jwt_required]
     file_parser = RequestParser(bundle_errors=True)
 
-    @ns_kyc.marshal_with(mschema)
+    # @ns_kyc.marshal_with(mschema)
     def post(self):
-        # if current_user.kyc_doc.status == 'Approved':
-        #     return {'message': 'You documents are approval'}
+        doc = current_user.kyc_doc.status if current_user.kyc_doc else abort(404)
+        if doc != 'Pending':
+            return {'message': 'You cant update kyc when not in pending state'},400
         KycUploadResource.file_parser.add_argument('file', required=True, location='files',
                                                    type=werkzeug.datastructures.FileStorage)
         args = KycUploadResource.file_parser.parse_args(strict=True)
@@ -117,10 +114,12 @@ class KycUploadResource(Resource):
         if kyc.file:
             delete_file(kyc.file)
         file = file_upload(args.file)
+        if file.get('error'):
+            return file, 400
         file.get('upload').save(file.get('full_path'))
+        args.update({'file': url_for('api.protected_dir', filename=file.get('filename'), _external=True)})
         kyc.file = file.get('filename')
-        kyc.save()
-        return kyc, 200
+        return kyc.save(**args), 200
 
 
 ns_kyc.add_resource(KYCResourceList, '/', endpoint='kycs')
